Fix : home에 있는 슬라이드 intro 페이지로 이동
