--- components/Home/index.js
@@ -1,4 +1,0 @@
-export * from './HomeP1';
-export * from './HomeP2';
-export * from './HomeP3';
-export * from './HomeP4';
