+++ components/Intro/index.js
@@ -1,0 +1,4 @@
+export * from './IntroP1';
+export * from './IntroP2';
+export * from './IntroP3';
+export * from './IntroP4';
